feat(scalars): detach url input from form
