+++ packages/design-system/src/scalars/components/url-field/index.tsx
@@ -1,1 +1,2 @@
-export * from "./url-field.js";
+export { UrlField, type UrlFieldProps } from "./url-field.js";
+export { UrlInput, type PlatformIcon, type UrlInputProps } from "./url-input.js";
